Implements 2 layer neural net with newton method
The first layer has variable size, depending on user input. The second layer is the output layer. When doing "backpropagation", from the second layer to the first layer, it computes the descent direction with a newton method.
It is not highly optimized, but it does it's job quite well.
It converges to a minimum faster than GD on the simple example problem, but not by very much. It needs to be tested on better datasets.
I pretend on creating another append, where newton method is used to compute descent direction on all layers.